Introduced react-router, new folder strucutre by domain entity and TSLINT
