+++ package.json
@@ -3,6 +3,12 @@
   "version": "0.1.0",
   "description": "A simple, hackable software for the international education industry",
   "main": "./dist/bundle.js",
+  "scripts": {
+    "build": "webpack --env prod",
+    "start": "webpack-dev-server --port 3000 --historyApiFallback true",
+    "static": "static -at 0.0.0.0 dist",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
   "dependencies": {
     "@shopify/polaris": "^1.3.1",
     "core-js": "^2.5.0",
@@ -13,16 +19,18 @@
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-redux": "^5.0.6",
+    "react-router-dom": "^4.1.2",
     "redux": "^3.7.2",
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
     "@types/html-webpack-plugin": "^2.28.0",
-    "@types/react-redux": "^5.0.4",
     "@types/node": "^8.0.9",
     "@types/react": "^15.0.24",
     "@types/react-dom": "^15.5.0",
     "@types/react-hot-loader": "^3.0.3",
+    "@types/react-redux": "^5.0.4",
+    "@types/react-router-dom": "^4.0.7",
     "@types/redux": "^3.6.0",
     "@types/webpack": "^3.0.2",
     "@types/webpack-env": "^1.13.0",
@@ -35,15 +43,11 @@
     "source-map-loader": "^0.2.1",
     "style-loader": "^0.18.2",
     "ts-node": "^3.2.0",
+    "tslint": "^5.6.0",
+    "tslint-react": "^3.2.0",
     "typescript": "^2.3.3",
     "webpack": "^2.6.0",
     "webpack-dev-server": "^2.4.5"
-  },
-  "scripts": {
-    "build": "webpack --env prod",
-    "start": "webpack-dev-server --port 3000",
-    "static": "static -at 0.0.0.0 dist",
-    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
