--- src/App.css
@@ -1,15 +1,0 @@
-.foo{
-    background: #ccc;
-}
-
-.floating-button{
-  width: 55px;
-  height: 55px;
-  border-radius: 50%;
-  background: #db4437;
-  position: fixed;
-  bottom: 30px;
-  right: 30px;
-  cursor: pointer;
-  box-shadow: 0px 2px 5px #666;
-}
