+++ src/flux/habooProvider.ts
@@ -1,4 +1,4 @@
-import { config } from '../config/config.dev';
+import { config } from '../config';
 import HabooSdk from '@haboo/haboo-sdk';
 
 export default new HabooSdk(config.apiUrl, config.publicKey);
