+++ src/index.tsx
@@ -1,27 +1,27 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { AppContainer } from "react-hot-loader";
-import App from "./App";
+import App from "./app/App";
 import '@shopify/polaris/styles.scss';
 
-const rootEl = document.getElementById("root");
+const rootElement = document.getElementById("root");
 ReactDOM.render(
   <AppContainer>
     <App />
   </AppContainer>,
-  rootEl
+  rootElement
 );
 
 // Hot Module Replacement API
 if (module.hot) {
-  module.hot.accept("./App", () => {
-    const NextApp = require<{default: typeof App}>("./App").default;
+  module.hot.accept("./app/App", () => {
+    const NextApp = require<{default: typeof App}>("./app/App").default;
     ReactDOM.render(
       <AppContainer>
         <NextApp />
       </AppContainer>
       ,
-      rootEl
+      rootElement
     );
   });
 }
